Chats: use tanstack virtual for chat messages
