+++ package.json
@@ -100,6 +100,8 @@
 		"@sentry/browser": "^7.39.0",
 		"@sentry/cli": "^2.18.1",
 		"@sindresorhus/slugify": "^2.2.1",
+		"@tanstack/react-query": "^4.40.1",
+		"@tanstack/react-virtual": "^3.13.12",
 		"@types/emoji-mart": "^3.0.13",
 		"@types/react-transition-group": "^4.4.12",
 		"@viz-js/viz": "^3.2.4",
